Extract methods from I18n::JS to exporter, clarify public interface.
